US334060: Sorted packages and renamed classes
