+++ checkout/access-checkout-android/src/main/java/com/worldpay/access/checkout/client/AccessCheckoutClient.kt
@@ -1,0 +1,7 @@
+package com.worldpay.access.checkout.client
+
+interface AccessCheckoutClient {
+
+    fun generateSession(cardDetails: CardDetails, sessionTypes: List<SessionType>)
+
+}
