+++ checkout/access-checkout-android/src/main/java/com/worldpay/access/checkout/client/SessionType.kt
@@ -1,0 +1,8 @@
+package com.worldpay.access.checkout.client
+
+enum class SessionType {
+
+    PAYMENTS_CVC_SESSION,
+    VERIFIED_TOKEN_SESSION
+
+}
