--- checkout/access-checkout-android/src/main/java/com/worldpay/access/checkout/client/checkout/CheckoutClient.kt
@@ -1,10 +1,0 @@
-package com.worldpay.access.checkout.client.checkout
-
-import com.worldpay.access.checkout.client.card.CardDetails
-import com.worldpay.access.checkout.client.token.TokenRequest
-
-interface CheckoutClient {
-
-    fun generateSession(cardDetails: CardDetails, tokenRequests: List<TokenRequest>)
-
-}
