--- checkout/access-checkout-android/src/main/java/com/worldpay/access/checkout/client/token/TokenRequest.kt
@@ -1,8 +1,0 @@
-package com.worldpay.access.checkout.client.token
-
-enum class TokenRequest {
-
-    SESSION_TOKEN,
-    VERIFIED_TOKEN
-
-}
