+++ checkout/access-checkout-android/src/main/java/com/worldpay/access/checkout/session/request/SessionRequestHandlerFactory.kt
@@ -1,0 +1,21 @@
+package com.worldpay.access.checkout.session.request
+
+import com.worldpay.access.checkout.session.request.handlers.PaymentsCvcSessionRequestHandler
+import com.worldpay.access.checkout.session.request.handlers.VerifiedTokensSessionRequestHandler
+
+internal class SessionRequestHandlerFactory(private val sessionRequestHandlerConfig: SessionRequestHandlerConfig) {
+
+    private val handlers = listOf(
+        VerifiedTokensSessionRequestHandler(
+            sessionRequestHandlerConfig
+        ),
+        PaymentsCvcSessionRequestHandler(
+            sessionRequestHandlerConfig
+        )
+    )
+
+    fun getTokenHandlers(): List<SessionRequestHandler> {
+        return handlers
+    }
+
+}
